--- highway_configs.py
@@ -1,26 +1,0 @@
-kinematic_highway_config = {
-    "action": {"type": "DiscreteMetaAction"},
-    "centering_position": [0.3, 0.5],
-    "collision_reward": -1,
-    "controlled_vehicles": 1,
-    "duration": 100,
-    "ego_spacing": 2,
-    "initial_lane_id": None,
-    "lanes_count": 2,
-    "manual_control": False,
-    'observation': {'type': 'Kinematics'},
-    "offroad_terminal": True,
-    "offscreen_rendering": True,
-    "other_vehicles_type": "highway_env.vehicle.behavior.IDMVehicle",
-    "policy_frequency": 1,
-    "real_time_rendering": False,
-    "render_agent": True,
-    "reward_speed_range": [20, 30],
-    "scaling": 5.5,
-    "screen_height": 150,
-    "screen_width": 600,
-    "show_trajectories": False,
-    "simulation_frequency": 15,
-    "vehicles_count": 50,
-    "vehicles_density": 1,
-}
